requirements.txt
django==4.2
djangorestframework==3.14.0
markdown
pydantic==2.1.1
psycopg[binary]==3.1.10
